some pkg/file nits
