+++ src/test/scala/org/hammerlab/magic/rdd/batch/MapReduceTest.scala
@@ -1,13 +1,11 @@
 package org.hammerlab.magic.rdd.batch
 
 import org.apache.spark.HashPartitioner
-import org.apache.spark.batch._
-import org.apache.spark.rdd.ParallelCollectionRDD
-
+import org.apache.spark.batch.{ MapRDD, ReduceRDD }
 import org.hammerlab.magic.rdd.batch.implicits._
 import org.hammerlab.spark.test.suite.SparkSuite
 
-class BatchMapReduceTest extends SparkSuite {
+class MapReduceTest extends SparkSuite {
 
   test("invalid batch size <= 0") {
     val rdd = sc.parallelize(0 until 10, 8)
@@ -35,25 +33,25 @@ class BatchMapReduceTest extends SparkSuite {
   test("return batch reduce RDD, if number of partitions is larger than batch size") {
     val rdd = sc.parallelize(0 until 10, 8)
     val res1 = rdd.batch(4)
-    res1.isInstanceOf[BatchReduceRDD[_]] should be (true)
+    res1.isInstanceOf[ReduceRDD[_]] should be (true)
     // we do allow to use batch size of 1 - one partition per stage
     val res2 = rdd.batch(1)
-    res2.isInstanceOf[BatchReduceRDD[_]] should be (true)
+    res2.isInstanceOf[ReduceRDD[_]] should be (true)
   }
 
   test("batch map RDD should have same hash partitioner of original partitions") {
     val rdd = sc.parallelize(0 until 10, 8)
-    val batch1 = new BatchMapRDD(rdd, None, rdd.partitions)
+    val batch1 = new MapRDD(rdd, None, rdd.partitions)
     batch1.part.isInstanceOf[HashPartitioner] should be (true)
-    val batch2 = new BatchMapRDD(rdd, Some(batch1), rdd.partitions)
+    val batch2 = new MapRDD(rdd, Some(batch1), rdd.partitions)
     batch2.part should be (batch1.part)
   }
 
   test("batch map RDD should return correct number of dependencies") {
     val rdd = sc.parallelize(0 until 10, 8)
-    val batch1 = new BatchMapRDD(rdd, None, rdd.partitions)
-    val batch2 = new BatchMapRDD(rdd, Some(batch1), rdd.partitions)
-    val batch3 = new BatchMapRDD(rdd, Some(batch2), rdd.partitions)
+    val batch1 = new MapRDD(rdd, None, rdd.partitions)
+    val batch2 = new MapRDD(rdd, Some(batch1), rdd.partitions)
+    val batch3 = new MapRDD(rdd, Some(batch2), rdd.partitions)
     batch1.getBatchDependencies should be (Seq.empty)
     batch2.getBatchDependencies should be (Seq(batch1))
     batch3.getBatchDependencies should be (Seq(batch1, batch2))
@@ -61,20 +59,20 @@ class BatchMapReduceTest extends SparkSuite {
 
   test("batch reduce RDD should fail if original partitions are reused") {
     val rdd = sc.parallelize(0 until 10, 8)
-    val batch1 = new BatchMapRDD(rdd, None, rdd.partitions)
-    val batch2 = new BatchMapRDD(rdd, Some(batch1), rdd.partitions)
+    val batch1 = new MapRDD(rdd, None, rdd.partitions)
+    val batch2 = new MapRDD(rdd, Some(batch1), rdd.partitions)
     val err = intercept[IllegalStateException] {
-      new BatchReduceRDD(batch2)
+      new ReduceRDD(batch2)
     }
     assert(err.getMessage.contains(s"Map-side RDD ${batch2.id} contains duplicate partition"))
   }
 
   test("batch reduce RDD should fail if original partitions are different from partition map") {
     val rdd = sc.parallelize(0 until 10, 8)
-    val batch1 = new BatchMapRDD(rdd, None, rdd.partitions.take(3))
-    val batch2 = new BatchMapRDD(rdd, Some(batch1), rdd.partitions.drop(4))
+    val batch1 = new MapRDD(rdd, None, rdd.partitions.take(3))
+    val batch2 = new MapRDD(rdd, Some(batch1), rdd.partitions.drop(4))
     val err = intercept[IllegalStateException] {
-      new BatchReduceRDD(batch2)
+      new ReduceRDD(batch2)
     }
     assert(err.getMessage.contains(
       "Partition-dependency map has 7 partitions, but RDD should have 8 partitions"))
@@ -82,9 +80,9 @@ class BatchMapReduceTest extends SparkSuite {
 
   test("batch reduce RDD should return same partitions as original RDD") {
     val rdd = sc.parallelize(0 until 10, 8)
-    val batch1 = new BatchMapRDD(rdd, None, rdd.partitions.take(3))
-    val batch2 = new BatchMapRDD(rdd, Some(batch1), rdd.partitions.drop(3))
-    val res = new BatchReduceRDD(batch2)
+    val batch1 = new MapRDD(rdd, None, rdd.partitions.take(3))
+    val batch2 = new MapRDD(rdd, Some(batch1), rdd.partitions.drop(3))
+    val res = new ReduceRDD(batch2)
     res.partitions should be (rdd.partitions)
   }
